+++ pinterest/delivery/handleFuncs.go
@@ -1,27 +1,26 @@
-package handls
+package delivery
 
 import (
-	"2019_2_Solar/pkg/structs"
+	"2019_2_Solar/pinterest"
+	"2019_2_Solar/pinterest/usecase"
+	"2019_2_Solar/pkg/models"
 	"log"
 	"net/http"
 	"sync"
 )
 
-var handlers = Handlers{
-	Users:    make([]structs.User, 0),
-	Sessions: make([]structs.UserSession, 0),
-	Mu:       &sync.Mutex{},
+type Handlers struct {
+	PUsecase pinterest.Usecase
+	Users    []models.User
+	Sessions []models.UserSession
+	Mu       *sync.Mutex
 }
 
-func CORSMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "*")
-		w.Header().Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
-		w.Header().Set("Access-Control-Allow-Headers", "Accept, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization")
-		w.Header().Set("Access-Control-Allow-Origin", "http://solar-env.v2zxh2s3me.us-east-2.elasticbeanstalk.com")
-		w.Header().Set("Access-Control-Allow-Credentials", "true")
-		next.ServeHTTP(w, r)
-	})
+var handlers = Handlers{
+	PUsecase: &usecase.PinterestUseCase{},
+	Users:    make([]models.User, 0),
+	Sessions: make([]models.UserSession, 0),
+	Mu:       &sync.Mutex{},
 }
 
 func HandleRoot(w http.ResponseWriter, r *http.Request) {
