+++ pinterest/repository.go
@@ -1,0 +1,4 @@
+package pinterest
+
+type repository interface {
+}
