+++ pinterest/repository/pinterest_repository.go
@@ -1,0 +1,3 @@
+package repository
+
+// Здесь будут методы для работы с БД
