+++ simple-server.go
@@ -1,18 +1,19 @@
 package main
 
 import (
-	"2019_2_Solar/pkg/handls"
+	"2019_2_Solar/pinterest/delivery"
+	"2019_2_Solar/pkg/middlewares"
 	"net/http"
 )
 
 func main() {
-	http.Handle("/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleRoot)))
-	http.Handle("/users/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleUsers)))
-	http.Handle("/registration/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleRegistration)))
-	http.Handle("/login/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleLogin)))
-	http.Handle("/logout/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleLogout)))
-	http.Handle("/profile/data", handls.CORSMiddleware(http.HandlerFunc(handls.HandleProfileData)))
-	http.Handle("/profile/picture", handls.CORSMiddleware(http.HandlerFunc(handls.HandleProfilePicture)))
+	http.Handle("/", middlewares.CORSMiddleware(http.HandlerFunc(delivery.HandleRoot)))
+	http.Handle("/users/", middlewares.CORSMiddleware(http.HandlerFunc(delivery.HandleUsers)))
+	http.Handle("/registration/", middlewares.CORSMiddleware(http.HandlerFunc(delivery.HandleRegistration)))
+	http.Handle("/login/", middlewares.CORSMiddleware(http.HandlerFunc(delivery.HandleLogin)))
+	http.Handle("/logout/", middlewares.CORSMiddleware(http.HandlerFunc(delivery.HandleLogout)))
+	http.Handle("/profile/data", middlewares.CORSMiddleware(http.HandlerFunc(delivery.HandleProfileData)))
+	http.Handle("/profile/picture", middlewares.CORSMiddleware(http.HandlerFunc(delivery.HandleProfilePicture)))
 
 	http.ListenAndServe(":8080", nil)
 }
